Add `groupBy` feature
To enable grouping of items of a list, provide `groupBy` query,
e.g. `http://brainfock.org/projects/?groupBy=workspace`.

To actually be able to use property for grouping, each item in list has
to have object under property name of same value, e.g. `workspace`,
and this object is required to have `id` and `name` or `summary` values.
Out of the box we are able to group by any topic relation, including but
not limited to parent, context, type, stage, assigee etc.